fix revert
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/plugins/XMLSpider@24221 67a373e5-eb02-0410-a15c-ee090a768436

+++ XMLSpider.java
@@ -1224,7 +1224,7 @@ public class XMLSpider implements FredPlugin, FredPluginHTTP, FredPluginThreadle
 			startSomeRequests();
 		}
 
-		Long lastPosition = null;
+		Integer lastPosition = null;
 		
 		public void onText(String s, String type, URI baseURI){
 			Logger.debug(this, "onText on " + page.id + " (" + baseURI + ")");
